ensure the DDG min-length defaults are respected in length checks
