+++ packages/password/lib/constants.js
@@ -1,0 +1,11 @@
+const MIN_LENGTH = 20
+const MAX_LENGTH = 30
+const DEFAULT_PASSWORD_RULES = `minlength: ${MIN_LENGTH}; maxlength: ${MAX_LENGTH};`
+
+const constants = {
+    MIN_LENGTH,
+    MAX_LENGTH,
+    DEFAULT_PASSWORD_RULES
+}
+
+module.exports.constants = constants
